requirements.txt
autopep8==1.4.2
Django==2.1
django-watson==1.5.2  
djangorestframework==3.8.2
pep8==1.7.1
pip==18.0   
postgis==1.0.4  
psycopg2==2.7.5  
psycopg2-binary==2.7.5  
pytz==2018.5
setuptools==40.5.0
djangorestframework-gis==0.13